chore: separate views and viewModels
